Maps integration completed

+++ FlightBooking/app/src/main/java/model/Flight.java
@@ -21,10 +21,10 @@ public class Flight implements Parcelable {
     private String time1;
     private String time2;
     private String duration;
-    private int townFromLatitude;
-    private int townToLatitude;
-    private int townFromLongitude;
-    private int townToLongitude;
+    private Double townFromLatitude;
+    private Double townToLatitude;
+    private Double townFromLongitude;
+    private Double townToLongitude;
 
 
     public Flight() {
@@ -45,14 +45,14 @@ public class Flight implements Parcelable {
         this.time1 = in.readString();
         this.time2 = in.readString();
         this.duration = in.readString();
-        this.townFromLatitude = in.readInt();
-        this.townToLatitude = in.readInt();
-        this.townFromLongitude = in.readInt();
-        this.townToLongitude = in.readInt();
+        this.townFromLatitude = in.readDouble();
+        this.townToLatitude = in.readDouble();
+        this.townFromLongitude = in.readDouble();
+        this.townToLongitude = in.readDouble();
 
     }
 
-    public Flight(int id, String townFrom, String townTo, String townFromMark, String townToMark, String price, String company, String date1, String date2, String time1, String time2, String duration, int townFromLatitude, int townToLatitude, int townFromLongitude, int townToLongitude) {
+    public Flight(int id, String townFrom, String townTo, String townFromMark, String townToMark, String price, String company, String date1, String date2, String time1, String time2, String duration, Double townFromLatitude, Double townToLatitude, Double townFromLongitude, Double townToLongitude) {
         this.id = id;
         this.townFrom = townFrom;
         this.townTo = townTo;
@@ -167,35 +167,35 @@ public class Flight implements Parcelable {
         this.duration = duration;
     }
 
-    public int getTownFromLatitude() {
+    public Double getTownFromLatitude() {
         return townFromLatitude;
     }
 
-    public void setTownFromLatitude(int townFromLatitude) {
+    public void setTownFromLatitude(Double townFromLatitude) {
         this.townFromLatitude = townFromLatitude;
     }
 
-    public int getTownToLatitude() {
+    public Double getTownToLatitude() {
         return townToLatitude;
     }
 
-    public void setTownToLatitude(int townToLatitude) {
+    public void setTownToLatitude(Double townToLatitude) {
         this.townToLatitude = townToLatitude;
     }
 
-    public int getTownFromLongitude() {
+    public Double getTownFromLongitude() {
         return townFromLongitude;
     }
 
-    public void setTownFromLongitude(int townFromLongitude) {
+    public void setTownFromLongitude(Double townFromLongitude) {
         this.townFromLongitude = townFromLongitude;
     }
 
-    public int getTownToLongitude() {
+    public Double getTownToLongitude() {
         return townToLongitude;
     }
 
-    public void setTownToLongitude(int townToLongitude) {
+    public void setTownToLongitude(Double townToLongitude) {
         this.townToLongitude = townToLongitude;
     }
